S-14d extract approvals to web/modules/approvals.js
Moves loadApprovalsPage, filter, pagination and _resolveApproval (~208 lines)
out of app.js into a self-contained ES module with DI-free initApprovals().
Removes escapeAttr from app.js (last use was in the approvals section).
Updates approvals-ui-contract.test.ts to assert against the module file.
Build + 2887 vitest green + headless smoke (0 pageerror, tbody/filter/pagination present).
